refactor: reorganize core.py into separate services
