+++ ads/tests/test_search.py
@@ -1,14 +1,114 @@
 """
-Tests for SearchQuery and it's query proxy class.
+Tests for the search interface
 """
 import sys
 import unittest
 import requests
+from mock import patch
+import six
 
 from .mocks import MockSolrResponse
 
-from ads.core import SearchQuery, query, SolrResponse, APIResponse
+from ads.search import SearchQuery, SolrResponse, APIResponse, Article
 from ads.exceptions import APIResponseError, SolrResponseParseError
+from ads.config import SEARCH_URL
+
+
+class TestArticle(unittest.TestCase):
+    """
+    Test the Article object
+    """
+
+    def setUp(self):
+        """
+        Create a test Article instance
+        """
+        self.article = Article(
+            bibcode="2013A&A...552A.143S",
+            year=2013,
+            first_author="Sudilovsky, V.",
+            author=[
+                "Sudilovsky, V",
+                "Greiner, J",
+                "Rau, A",
+                "Salvato, M",
+                "Savaglio, S",
+                "Vergani, S",
+                "Schady, P",
+                "Elliott, J",
+                "Kruehler, T",
+                "Kann, D",
+                "Klose, S",
+                "Rossi, A",
+                "Filgas, R",
+                "Schmidl, S"
+            ],
+        )
+
+    def test_equals(self):
+        """
+        the __eq__ method should compare bibcodes, and raise if bibcode isn't
+        defined
+        """
+        self.assertNotEqual(Article(bibcode="Not the same"), self.article)
+        self.assertEqual(Article(bibcode="2013A&A...552A.143S"), self.article)
+        with self.assertRaises(TypeError):
+            Article() == self.article
+
+    def test_init(self):
+        """
+        after init ._raw should be a dict containing a subset of all
+        class attributes
+        """
+        for key, value in six.iteritems(self.article._raw):
+            self.assertEqual(
+                self.article.__getattribute__(key),
+                value,
+                msg="Instance attribute and _raw mismatch on {}".format(key)
+            )
+
+    def test_print_methods(self):
+        """
+        the class should return a user-friendly formatted identified when the
+        __str__ or __unicode__ methods are called
+        """
+        self.assertEqual(
+            '<Sudilovsky, V. et al. 2013, 2013A&A...552A.143S>',
+            self.article.__str__()
+        )
+        self.assertEqual(self.article.__unicode__(), self.article.__str__())
+
+    @patch('ads.search.Article._get_field')
+    def test_cached_properties(self, patched):
+        """
+        The underlying computation responsible for filling in cached properties
+        should be called once and only once per attribute.
+        In addition, this computation should only be called if the instance
+        attribute wasn't set elsewhere.
+        """
+        patched.return_value = "patched response"
+        self.assertEqual(self.article.bibcode, '2013A&A...552A.143S')
+        self.assertEqual(self.article.aff, 'patched response')
+        self.assertEqual(self.article.aff, 'patched response')
+        patched.assert_called_once_with('aff')
+
+    def test_get_field(self):
+        """
+        The helper method _get_field() should return the value of a field
+        based on its `id`
+        """
+        with self.assertRaisesRegexp(
+                APIResponseError,
+                "Cannot query an article without an id"):
+            self.article._get_field('read_count')
+
+        # Note that our mock solr server doesn't do anything with "q", and
+        # therefore will just return the first article in the hardcoded
+        # stubdata. We assume the live service will return the correct document.
+        self.article.id = 9535116
+        with MockSolrResponse(SEARCH_URL):
+            self.assertEqual(self.article.pubdate, '1971-10-00')
+            self.assertEqual(self.article.read_count, 0.0)
 
 
 class TestSearchQuery(unittest.TestCase):
@@ -130,8 +230,6 @@ class TestSolrResponse(unittest.TestCase):
         self.response.status_code = 500
         with self.assertRaises(APIResponseError):
             SolrResponse.load_http_response(self.response)
-
-
 
 if __name__ == '__main__':
     unittest.main(verbosity=2)
